% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareRandomEfectSD.R
\name{CompareRandomEfectSD}
\alias{CompareRandomEfectSD}
\title{Compare two mdoles in terms of the standard deviation of their Random effects}
\usage{
CompareRandomEfectSD(mod1 = NULL, mod2 = NUll, rnd = 1,
  percentage = TRUE)
}
\arguments{
\item{mod1}{object of \code{\link{class}} \code{inla} with random marginals.}

\item{mod2}{object of \code{\link{class}} \code{inla} with the same random marginals of mod1.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{vector}} if only one model with one random effect is evaluated; \code{\link{data.frame}} or \code{\link{list}} otherwise.
}
\description{
Compare two mdoles in terms of the standard deviation of their Random effects.
}
\examples{

}
