% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO.R
\name{CPO}
\alias{CPO}
\title{Conditional predictive ordinates}
\usage{
CPO(..., failure_cutoff = 0, decreasing = TRUE)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla}  with computed CPO's.}

\item{failure_cutoff}{value between 0 and 1. CPO's with a failure score greater than \code{failure_cutoff} are considered unreliable and are ignored.}

\item{decreasing}{logical. If \code{TRUE} (default), CPO's are displayed in decreasing order.}
}
\value{
\code{\link{matrix}} with the sum of the log of CPO and the proportion of failures of each models' CPO.
}
\description{
Sort and display the sum of the log of CPO's of INLA models.
}
\examples{
data(sp)

mod1 <- inla(aan ~ offset(log(eaan)) +
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,
             control.compute = list(cpo = TRUE))

mod2 <- inla(aan ~ shvn + offset(log(eaan))+
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,
             control.compute = list(cpo = TRUE))

CPO(mod1, mod2)

}
\references{
Blangiardo, M., Cameletti, M., Baio, G., & Rue, H. (2013). Spatial and spatio-temporal models with R-INLA. Spatial and spatio-temporal epidemiology, 7, 39-55.
}
