% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Cross-validated model performance}
\usage{
CV(..., k = NULL, test_prop = NULL, test_idx = NULL, rep = 1,
  seed = NULL, decreasing = FALSE, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla}  with computed fitted marginals.}

\item{k}{number of folds in k-fold cross-validation.}

\item{test_prop}{proportion of observations to compose the test set in a training/test split.}

\item{test_idx}{numeric index of observations to compose the test set in a training/test split.}

\item{rep}{number of repetitions of the cross-validation procedure.}

\item{seed}{number passed to \code{\link{set.seed}} to avoid variation in the data split invoked by \code{k} or \code{test_prop}.}

\item{decreasing}{logical. If \code{FALSE} (default), RMSE's are displayed in increasing order.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{vector}} with mean RMSE for each model.
}
\description{
Calculates the RMSE of INLA models using cross-validation.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.predictor=list(link = 1, compute = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.predictor=list(link = 1, compute = TRUE))
CV(mod1, mod2, test_prop = .1, rep = 5)
CV(mod1, mod2, k = 3, rep = 2)
}
