% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Deviance information criteria}
\usage{
DIC(mods, decreasing = F)
}
\arguments{
\item{mods}{\code{\link{list}} of objects of \code{\link{class}} \code{inla} with computed DIC.}

\item{decreasing}{logical. If \code{FALSE} (default), DIC's are displayed in increasing order.}
}
\value{
\code{\link{matrix}} with models' DIC.
}
\description{
Sort and display the Deviance information criteria of INLA models.
}
\examples{
data(sp)

mod1 <- inla(aan ~ offset(log(eaan)) +
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,
             control.compute = list(dic = TRUE))

mod2 <- inla(aan ~ shvn + offset(log(eaan))+
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,
             control.compute = list(dic = TRUE))

DIC(list(mod1 = mod1, mod2 = mod2))
}
\references{
Blangiardo, M., Cameletti, M., Baio, G., & Rue, H. (2013). Spatial and spatio-temporal models with R-INLA. Spatial and spatio-temporal epidemiology, 7, 39-55.
}

