% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplainedVariance.R
\name{ExplainedVariance}
\alias{ExplainedVariance}
\title{Variance explained by each variance component}
\usage{
ExplainedVariance(..., n = 10000, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with with one or more random effects.}

\item{n}{number of random values generated from hyperparameters.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{data.frame}} with the proportion of variance explained by each variance component.
}
\description{
Proportion of variance explained by each variance component.
}
\details{
The argument \code{scale.model} must be TRUE within \code{f} function.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj,
                     scale.model = TRUE),
             E = eaan,
             family = 'poisson', data = spn)

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj,
                            scale.model = TRUE),
             E = eaan,
             family = 'poisson', data = spn)

ExplainedVariance(mod1, mod2)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
