% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedExcess.R
\name{FittedExcess}
\alias{FittedExcess}
\title{Fitted values in excess}
\usage{
FittedExcess(..., cutoff = 1, rnd = 3)
}
\arguments{
\item{...}{object of \code{\link{class}} \code{inla} with fitted values.}

\item{cutoff}{value above which fittde values would be in excess.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{vector}} if only one model with fitted values is evaluated; \code{\link{data.frame}} or \code{\link{list}} otherwise.
}
\description{
@description Probability of fitted values being higher than a cutoff value.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn,
             control.predictor = list(compute = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn,
             control.predictor = list(compute = TRUE))

see <- RandomEffectsExcess(mod1, mod2, cutoff = 1)
summary(see)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
