% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffects.R
\name{FixedEffects}
\alias{FixedEffects}
\title{Fixed effects}
\usage{
FixedEffects(model = NULL, quantiles = c(0.025, 0.975), hpd = 0.95,
  sd = FALSE, expo = TRUE, rnd = 3)
}
\arguments{
\item{model}{object of \code{\link{class}} \code{inla} with fixed marginals.}

\item{quantiles}{\code{\link{numeric}} \code{\link{vector}} to indicate the quantiles to be computed.}

\item{sd}{logical. If \code{FALSE} (default), the standard deviation is not returned.}

\item{expo}{logical. If \code{TRUE} (default), summary statistics are expoentiated.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{hpd.}{Proportion of the distribution included in the highest density interval.}
}
\value{
\code{\link{matrix}} with summary statistics for relative risks or odds ratios.
}
\description{
Summary statistics of fixed marginals
}
\details{
Exponentiated fixed marginals represent relative risks for models with one of the following likelihoods: \code{poisson}, \code{zeroinflated.poisson.0}, \code{zeroinflated.poisson.1}, \code{zeroinflated.poisson.2}, \code{nbinomial}, \code{zeroinflated.nbinomial.0}, \code{zeroinflated.nbinomial.1}, \code{zeroinflated.nbinomial.2}. Exponentiated random marginals represent odds ratios for models with one of the followinglikelihoods: \code{binomial}, \code{zeroinflated.binomial.0}, \code{zeroinflated.binomial.1}.
}
\examples{
data(sp)
mod <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
            E = eaan, family = 'poisson', data = spn)
FixedEffects(mod)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
