% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/INLAOutputs-package.R
\docType{package}
\name{INLAOutputs-package}
\alias{INLAOutputs-package}
\title{The INLAOutputs Package}
\description{
Process Outputs form 'INLA' Package.
}
\details{
\tabular{ll}{
Package: \tab INLAOutputs\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2015-08-26\cr
Depends: \tab R (>= 3.2.1), INLA\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/oswaldosantos/INLAOutputs}\cr
Author: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
}
}
\keyword{package}

