% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prec2SD.R
\name{Prec2SD}
\alias{Prec2SD}
\title{Standard deviation of hyperparameters}
\usage{
Prec2SD(model, summarise = TRUE, sd = TRUE, quantiles = c(0.025,
  0.975), rnd = 3)
}
\arguments{
\item{model}{object of \code{\link{class}} \code{inla}.}

\item{summarise}{\code{\link{logical}}. If \code{TRUE} (default), the posterior mean of the standard deviation is returned.}

\item{quantiles}{\code{\link{numeric}} \code{\link{vector}} to indicate the quantiles to be computed.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{matrix}} with summary statistics.
}
\description{
Transformation of precision to standard deviation of hyperparameters.
}
\examples{
data(sp)
mod <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
            E = eaan, family = 'poisson', data = spn)
Prec2SD(mod)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
