% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root mean square error}
\usage{
RMSE(..., observed = NULL, decreasing = FALSE, na.rm = FALSE)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla}  with computed fitted marginals.}

\item{decreasing}{logical. If \code{FALSE} (default), DIC's are displayed in increasing order.}
}
\value{
na.rm \code{\link{logical}} if \code{FALSE}, NA's are note removed.
}
\description{
Calculates the RMSE of INLA models.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.predictor=list(link = 1, compute = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.predictor=list(link = 1, compute = TRUE))

}
