% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialEffectsExcess.R
\name{SpatialEffectsExcess}
\alias{SpatialEffectsExcess}
\title{Spatial effects in excess}
\usage{
SpatialEffectsExcess(..., rnd = 3)
}
\arguments{
\item{...}{object of \code{\link{class}} \code{inla} with outputs from a BYM model.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
\code{\link{matrix}} with as many rows as areas and as many columns as models in \code{mod}.
}
\description{
Probability of the area-specific relative risk (odds ratio) being > 1. Mean relative risk (odds ratio) is the reference category.
}
\details{
Spatial effects in excess are represented in terms of relative risks for models with one of the following likelihoods: \code{poisson}, \code{zeroinflated.poisson.0}, \code{zeroinflated.poisson.1}, \code{zeroinflated.poisson.2}, \code{nbinomial}, \code{zeroinflated.nbinomial.0}, \code{zeroinflated.nbinomial.1}, \code{zeroinflated.nbinomial.2}. Sptial effects in excess are represented in terms of odds ratios for models with one of the followinglikelihoods: \code{binomial}, \code{zeroinflated.binomial.0}, \code{zeroinflated.binomial.1}.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

see <- SpatialEffectsExcess(mod1, mod2)
summary(see)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
