% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSpatiallyExplained.R
\name{VarSpatiallyExplained}
\alias{VarSpatiallyExplained}
\title{Variance spatially explained}
\usage{
VarSpatiallyExplained(..., sims = 1000, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with outputs from a BYM model.}

\item{sims}{number of random values generated for each area, from the corresponding marginal posterior distribution of the unstructured residual component.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
Proportion of variance explained by the spatially structured component, for each model.
}
\description{
Proportion of variance explained by the spatially structured component.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

VarSpatiallyExplained(mod1, mod2)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
