% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO.R
\name{CPO}
\alias{CPO}
\title{Conditional predictive ordinates}
\usage{
CPO(..., failure_cutoff = 0, decreasing = TRUE)
}
\arguments{
\item{...}{models of \code{\link{class}} \code{inla} with computed CPO's.}

\item{failure_cutoff}{value between 0 and 1. CPO's with a failure score greater than \code{failure_cutoff} are considered unreliable and are ignored.}

\item{decreasing}{logical. If \code{TRUE} (default), CPO's are displayed in decreasing order (higher values indicate better prediction).}
}
\value{
\code{\link{matrix}} with the sum of the log of CPO's and the proportion of failures of each models' CPO.
}
\description{
Sort and display the sum of the log of CPO's of INLA models.
}
\details{
\code{control.compute = list(cpo = TRUE)} must be used within \code{inla} function.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.compute = list(cpo = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,  E = eaan,
             control.compute = list(cpo = TRUE))

CPO(mod1, mod2)

}
\references{
Blangiardo, M., Cameletti, M., Baio, G., & Rue, H. (2013). Spatial and spatio-temporal models with R-INLA. Spatial and spatio-temporal epidemiology, 7, 39-55.
}
