% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Deviance information criteria}
\usage{
DIC(..., decreasing = FALSE)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla}  with computed DIC.}

\item{decreasing}{logical. If \code{FALSE} (default), DIC's are displayed in increasing order.}
}
\value{
\code{\link{matrix}} with models' DIC, mean deviance and effective number of paramenters.
}
\description{
Sort and display the Deviance information criteria of INLA models.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.compute = list(dic = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.compute = list(dic = TRUE))

DIC(mod1, mod2)

}
\references{
Blangiardo, M., Cameletti, M., Baio, G., & Rue, H. (2013). Spatial and spatio-temporal models with R-INLA. Spatial and spatio-temporal epidemiology, 7, 39-55.
}
