% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAOutputs-package.R
\docType{package}
\name{INLAOutputs-package}
\alias{INLAOutputs-package}
\title{The INLAOutputs Package}
\description{
Process selected outputs from the 'INLA' package
}
\details{
\tabular{ll}{
Package: \tab INLAOutputs\cr
Type: \tab Package\cr
Version: \tab 1.4.10\cr
Date: \tab 2018-10-02\cr
Depends: \tab R (>= 3.4), INLA\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/oswaldosantos/INLAOutputs}\cr
Author: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
}
}
\keyword{package}
