% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredPValue.R
\name{PredPValue}
\alias{PredPValue}
\title{Posterior predictive p-value}
\usage{
PredPValue(..., observed = NULL, cutoff = c(0.1, 0.9), decreasing = FALSE)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with computed fitted marginals.}

\item{observed}{\code{\link{vector}} observed values.}

\item{cutoff}{\code{\link{vector}} with lower and upper values to define the tails of the p-values cumulative distribution. Default: c(0.1, 0.9).}

\item{decreasing}{logical. If \code{FALSE} (default), models are displayed in increasing order, according to the proportion of values in both tails (the first model has the best fit).}
}
\value{
\code{\link{list}}. The first element contains the proportion of p-values in the lower and upper tails. The second element contains the p-values.
}
\description{
Computes de probability Pr(predicted_i <= observed_i | observed).
}
\details{
\code{control.predictor = list(link = 1, compute = TRUE)} must be used within \code{inla} function.
}
\examples{
data(sp)

mod <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E= eaan,
             control.predictor = list(link = 1, compute = TRUE))

PredPValue(mod, observed = spn$aan)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
