% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredPValue.R
\name{PredPValue}
\alias{PredPValue}
\title{Posterior predictive p-value}
\usage{
PredPValue(mod, observed = NULL, cutoff = c(0.1, 0.9))
}
\arguments{
\item{mod}{object of \code{\link{class}} \code{inla}  with computed fitted marginal.}

\item{observed}{\code{\link{vector}} observed values.}

\item{cutoff}{\code{\link{vector}} with lower and upper values to define the tails of the p-values cumulative distribution. Default: c(0.1, 0.9).}
}
\value{
\code{\link{list}}. The first element contains the proportion of p-values in the lower and upper tails. The second element contains the p-values.
}
\description{
Computes de probability Pr(predicted_i <= observed_i | observed).
}
\examples{
data(sp)

mod <- inla(aan ~ offset(log(eaan)) +
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn,
             control.predictor=list(link = 1, compute = TRUE))

PredPValue(mod, spn$aan)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
