% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorityIndex.R
\name{PriorityIndex}
\alias{PriorityIndex}
\title{Priority index.}
\usage{
PriorityIndex(..., effect = NULL, cutoff = NULL, rnd = 1)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with fitted values or random marginals.}

\item{effect}{string indicating if the priority index should be calculated using fitted values \code{"fitted"} or random effects \code{"random"}.}

\item{cutoff}{value above which fitted values or random effects would be in excess. If \code{effect = "random"}, the cutoff is used for untransformed marginal posteriors, so \code{cutoff = 0} (default) is used to calculate the probability of relative risk (odds ratio) > 1.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{vector}} if only one model with fitted values is evaluated; \code{\link{data.frame}} or \code{\link{list}} otherwise.
}
\description{
@description Probability of fitted values being higher than a cutoff value.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn,
             control.predictor = list(compute = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn,
             control.predictor = list(compute = TRUE))

see <- RandomEffectsExcess(mod1, mod2, cutoff = 1)
}
