% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffects.R
\name{RandomEffects}
\alias{RandomEffects}
\title{Random effects}
\usage{
RandomEffects(..., expo = TRUE, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with random marginals. Only provide more than one object if the number of aggregation units is the same for all random marginals sets among all models (i.e. when the random model (iid, BYM, etc.) is the same for all objects).}

\item{expo}{logical. If \code{TRUE} (default), summary statistics are expoentiated.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
\code{\link{vector}} if only one model with one random effect is evaluated; \code{\link{data.frame}} or \code{\link{list}} otherwise.
}
\description{
Alternative display of computed random marginals.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)
summary(RandomEffects(mod1))

}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
