% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffectsExcess.R
\name{RandomEffectsExcess}
\alias{RandomEffectsExcess}
\title{Rnadom effects in excess}
\usage{
RandomEffectsExcess(..., cutoff = 0, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} with random marginals.}

\item{cutoff}{value above which random effects would be in excess. The cutoff is used for untransformed marginal posteriors, so \code{cutoff = 0} (default) is used to calculate the probability of relative risk (odds ratio) > 1.}

\item{rnd}{integer indicating the number of decimal places to be used.}
}
\value{
\code{\link{vector}} if only one model with one random effect is evaluated; \code{\link{data.frame}} or \code{\link{list}} otherwise.
}
\description{
Probability of random effects being higher than a cutoff value.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

see <- RandomEffectsExcess(mod1, mod2)
summary(see)
}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
