% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitIntercepts.R
\name{UnitIntercepts}
\alias{UnitIntercepts}
\title{Unit intercepts}
\usage{
UnitIntercepts(..., ran_marg = NULL, expo = TRUE, rnd = 3)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla} in random marginals.}

\item{ran_marg}{when objects passed to \code{...} have more than one random marginal and all have a common random marginal with the same ID, this ID can be specified as a string. \code{ran_marg} can be ignored for objects with only one random marginal.}

\item{expo}{logical. If \code{TRUE} (default), summary statistics are expoentiated.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
\code{\link{matrix}} with as many rows as areas and as many columns as models. Exponentiated values represent the posterior mean of unit-specific relative risks (odds ratio), compared to the mean relative risk (odds ratio).
}
\description{
Fixed intercept plus random marginals of each unit.
}
\examples{
head(spn)
mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn)

## Area-specific relative-risks.

summary(UnitIntercepts(mod1, expo = F))

lcs <- inla.make.lincombs(id = diag(nrow(spn)),
                          '(Intercept)' = rep(1, nrow(spn)))
mod2 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             E = eaan,
             family = 'poisson', data = spn,
             lincomb = lcs)
summary(mod2$summary.lincomb.derived[ , 2])

summary(mod2$summary.random[[1]][, 2] +
            as.numeric(mod2$summary.fixed[1]))

summary(mod1$summary.random[[1]][, 2] +
            as.numeric(mod1$summary.fixed[1]))

# SMR for comparison
summary(spn$aan / spn$eaan)
summary(UnitIntercepts(mod1))
summary(RandomEffects(mod1) * FixedEffects(mod1)[1, "mean"])

## Area-specific relative risks compaared to the mean relative risk.
summary(RandomEffects(mod1, expo = FALSE))
summary(mod2$summary.lincomb.derived[ , 2] - 
            as.numeric(mod2$summary.fixed[1]))

}
\references{
Blangiardo, Marta, and Michela Cameletti. Spatial and Spatio-temporal Bayesian Models with R-INLA. John Wiley & Sons, 2015.
}
