% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VarSpatiallyExplained.R
\name{VarSpatiallyExplained}
\alias{VarSpatiallyExplained}
\title{Variance spatially explained}
\usage{
VarSpatiallyExplained(mod = NULL, sims = 1000, rnd = 3)
}
\arguments{
\item{mod}{object of \code{\link{class}} \code{inla} or list of objects of \code{\link{class}} \code{inla}. Objects must have outputs from a BYM model.}

\item{sims}{number of random values generated for each area, from the corresponding marginal posterior distribution of the unstructured residual component.}

\item{rnd}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
Proportion of variance explained by the spatially structured component, for each of the models in \code{mods}.
}
\description{
Proportion of variance explained by the spatially structured component.
}
\examples{
data(sp)

mod1 <- inla(aan ~ offset(log(eaan)) +
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn)

mod2 <- inla(aan ~ shvn + offset(log(eaan))+
                 f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn)

models <- list(mod1, mod2)

VarSpatiallyExplained(models)
}
\references{
Blangiardo, M., Cameletti, M., Baio, G., & Rue, H. (2013). Spatial and spatio-temporal models with R-INLA. Spatial and spatio-temporal epidemiology, 7, 39-55.
}

