% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAIC.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe-Akaike information criteria}
\usage{
WAIC(..., decreasing = FALSE)
}
\arguments{
\item{...}{objects of \code{\link{class}} \code{inla}  with computed WAIC.}

\item{decreasing}{logical. If \code{FALSE} (default), WAIC's are displayed in increasing order.}
}
\value{
\code{\link{matrix}} with models' WAIC.
}
\description{
Sort and display the Watanabe-Akaike information criteria of INLA models.
}
\details{
\code{control.compute = list(waic = TRUE)} must be used within \code{inla} function.
}
\examples{
data(sp)

mod1 <- inla(aan ~ f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.compute = list(waic = TRUE))

mod2 <- inla(aan ~ shvn + f(id, model = 'bym', graph = sp.adj),
             family = 'poisson', data = spn, E = eaan,
             control.compute = list(waic = TRUE))

WAIC(mod1, mod2)

}
\references{
Gelman, Andrew, Jessica Hwang, and Aki Vehtari. "Understanding predictive information criteria for Bayesian models." Statistics and Computing 24.6 (2014): 997-1016.
}
